Introduce stream_utils.py
