+++ target_snowflake/__init__.py
@@ -8,16 +8,13 @@ import os
 import sys
 import copy
 
-from datetime import datetime
-from decimal import Decimal
 from typing import Dict, List
-from dateutil import parser
-from dateutil.parser import ParserError
 from joblib import Parallel, delayed, parallel_backend
 from jsonschema import Draft7Validator, FormatChecker
 from singer import get_logger
 
 import target_snowflake.file_formats.csv as csv
+import target_snowflake.stream_utils as stream_utils
 
 from target_snowflake.db_sync import DbSync
 from target_snowflake.exceptions import (
@@ -34,23 +31,6 @@ logging.getLogger('snowflake.connector').setLevel(logging.WARNING)
 DEFAULT_BATCH_SIZE_ROWS = 100000
 DEFAULT_PARALLELISM = 0  # 0 The number of threads used to flush tables
 DEFAULT_MAX_PARALLELISM = 16  # Don't use more than this number of threads by default when flushing streams in parallel
-
-# max timestamp/datetime supported in SF, used to reset all invalid dates that are beyond this value
-MAX_TIMESTAMP = '9999-12-31 23:59:59.999999'
-
-# max time supported in SF, used to reset all invalid times that are beyond this value
-MAX_TIME = '23:59:59.999999'
-
-
-def float_to_decimal(value):
-    """Walk the given data structure and turn all instances of float into double."""
-    if isinstance(value, float):
-        return Decimal(str(value))
-    if isinstance(value, list):
-        return [float_to_decimal(child) for child in value]
-    if isinstance(value, dict):
-        return {k: float_to_decimal(v) for k, v in value.items()}
-    return value
 
 
 def add_metadata_columns_to_schema(schema_message):
@@ -69,18 +49,6 @@ def add_metadata_columns_to_schema(schema_message):
     return extended_schema_message
 
 
-def add_metadata_values_to_record(record_message):
-    """Populate metadata _sdc columns from incoming record message
-    The location of the required attributes are fixed in the stream
-    """
-    extended_record = record_message['record']
-    extended_record['_sdc_extracted_at'] = record_message.get('time_extracted')
-    extended_record['_sdc_batched_at'] = datetime.now().isoformat()
-    extended_record['_sdc_deleted_at'] = record_message.get('record', {}).get('_sdc_deleted_at')
-
-    return extended_record
-
-
 def emit_state(state):
     """Print state to stdout"""
     if state is not None:
@@ -88,22 +56,6 @@ def emit_state(state):
         LOGGER.info('Emitting state %s', line)
         sys.stdout.write("{}\n".format(line))
         sys.stdout.flush()
-
-
-def get_schema_names_from_config(config):
-    """Get list of target schema name from config"""
-    default_target_schema = config.get('default_target_schema')
-    schema_mapping = config.get('schema_mapping', {})
-    schema_names = []
-
-    if default_target_schema:
-        schema_names.append(default_target_schema)
-
-    if schema_mapping:
-        for target in schema_mapping.values():
-            schema_names.append(target.get('target_schema'))
-
-    return schema_names
 
 
 def load_table_cache(config):
@@ -114,45 +66,9 @@ def load_table_cache(config):
 
         db = DbSync(config) # pylint: disable=invalid-name
         table_cache = db.get_table_columns(
-            table_schemas=get_schema_names_from_config(config))
+            table_schemas=stream_utils.get_schema_names_from_config(config))
 
     return table_cache
-
-
-def adjust_timestamps_in_record(record: Dict, schema: Dict) -> None:
-    """
-    Goes through every field that is of type date/datetime/time and if its value is out of range,
-    resets it to MAX value accordingly
-    Args:
-        record: record containing properties and values
-        schema: json schema that has types of each property
-    """
-
-    # creating this internal function to avoid duplicating code and too many nested blocks.
-    def reset_new_value(record: Dict, key: str, _format: str):
-        if not isinstance(record[key], str):
-            raise UnexpectedValueTypeException(
-                f'Value {record[key]} of key "{key}" is not a string.')
-
-        try:
-            parser.parse(record[key])
-        except ParserError:
-            LOGGER.warning('Parsing the %s "%s" in key "%s" has failed, thus defaulting to max '
-                           'acceptable value of %s in Snowflake', _format, record[key], key, _format)
-            record[key] = MAX_TIMESTAMP if _format != 'time' else MAX_TIME
-
-    # traverse the schema looking for properties of some date type
-    for key, value in record.items():
-        if value is not None and key in schema['properties']:
-            if 'anyOf' in schema['properties'][key]:
-                for type_dict in schema['properties'][key]['anyOf']:
-                    if 'string' in type_dict['type'] and type_dict.get('format', None) in {'date-time', 'time', 'date'}:
-                        reset_new_value(record, key, type_dict['format'])
-                        break
-            else:
-                if 'string' in schema['properties'][key]['type'] and \
-                        schema['properties'][key].get('format', None) in {'date-time', 'time', 'date'}:
-                    reset_new_value(record, key, schema['properties'][key]['format'])
 
 
 # pylint: disable=too-many-locals,too-many-branches,too-many-statements,invalid-name
@@ -192,12 +108,12 @@ def persist_lines(config, lines, table_cache=None) -> None:
             # Get schema for this record's stream
             stream = o['stream']
 
-            adjust_timestamps_in_record(o['record'], schemas[stream])
+            stream_utils.adjust_timestamps_in_record(o['record'], schemas[stream])
 
             # Validate record
             if config.get('validate_records'):
                 try:
-                    validators[stream].validate(float_to_decimal(o['record']))
+                    validators[stream].validate(stream_utils.float_to_decimal(o['record']))
                 except Exception as ex:
                     if type(ex).__name__ == "InvalidOperation":
                         raise InvalidValidationOperationException(
@@ -221,7 +137,7 @@ def persist_lines(config, lines, table_cache=None) -> None:
 
             # append record
             if config.get('add_metadata_columns') or config.get('hard_delete'):
-                records_to_load[stream][primary_key_string] = add_metadata_values_to_record(o)
+                records_to_load[stream][primary_key_string] = stream_utils.add_metadata_values_to_record(o)
             else:
                 records_to_load[stream][primary_key_string] = o['record']
 
@@ -250,7 +166,7 @@ def persist_lines(config, lines, table_cache=None) -> None:
                 raise Exception(f"Line is missing required key 'stream': {line}")
 
             stream = o['stream']
-            new_schema = float_to_decimal(o['schema'])
+            new_schema = stream_utils.float_to_decimal(o['schema'])
 
             # Update and flush only if the the schema is new or different than
             # the previously used version of the schema
